// packages/sdk/src/srd/5e-fastapi.d.ts
export interface components {
    schemas: {
// ...
        ItemEquipamento: {
            /**
             * Nome
             * @description Nome do item, como no livro (ex.: "Algibeira")
             */
            nome: string;
            /**
             * Quantidade
             * @description Quantidade do item
             * @default 1
             */
            quantidade: number;
        };
// ...
        OpcaoEquipamento: {
            /**
             * Opcao
             * @description Identificador da opção, como no livro (ex.: "A")
             */
            opcao: string;
            /**
             * Ouro
             * @description Peças de ouro (PO) da opção
             * @default 0
             */
            ouro: number;
            /**
             * Ferramentas
             * @description Ferramentas da opção
             */
            ferramentas?: components["schemas"]["ItemEquipamento"][];
            /**
             * Armas
             * @description Armas da opção
             */
            armas?: components["schemas"]["ItemEquipamento"][];
            /**
             * Outros
             * @description Demais itens da opção (ex.: algibeiras, roupas)
             */
            outros?: components["schemas"]["ItemEquipamento"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}